Moved Dialog component to a base component directory
