+++ src/components/EditBorderDialog.vue
@@ -26,7 +26,7 @@
 </template>
 
 <script>
-import Dialog from "@/components/Dialog.vue";
+import Dialog from "@/components/base/Dialog.vue";
 export default {
     components: {
         Dialog
